Bug fixes
Removed print statements.
Fixed a bug where loading a puzzle from a key after already having a game in progress didn't reset the player stats.
Added a prompt to the user to save their game when using /back in a current game.
Added a prompt to the user to save their game when using /exit from an active game.