Changes to database
Changed database to be more dynamic.
Puzzle can be any 7 letter pangram and the required letter can be any character in that pangram.
Added required letter, shuffle, and max points of a puzzle to the save / load system.
Added a function to calculate points of words on the fly. If a word is 4 letters long, 1 point. If a word has all unique characters, it's the length of the word * 2. Otherwise, it's the length of the word.